refactor: TestResultPage 컴포넌트 분리 및 상수 정리
