+++ src/features/test-result/ui/TestResultPage.tsx
@@ -1,12 +1,8 @@
 import { useState } from "react";
-import { AnimatePresence, motion } from "framer-motion";
 import {
   Asset,
-  BottomCTA,
   Button,
   ListHeader,
-  ListRow,
-  Post,
   Result,
   Skeleton,
   Tab,
@@ -17,132 +13,21 @@ import { adaptive } from "@toss/tds-colors";
 import { ResultTabContent } from "./ResultTabContent";
 import { QuestionTabContent } from "./QuestionTabContent";
 import { DownloadSheet } from "./DownloadSheet";
+import { ReviewGuideAccordion } from "./ReviewGuideAccordion";
+import { QuestionPreviewOverlay } from "./QuestionPreviewOverlay";
 import { mapReportItemToQuestionResult } from "../model/mappers";
+import { STATUS_BADGE } from "../model/constants";
 import { useGetReportQuery } from "@/shared/api/report";
 import type { TestStatus } from "@/shared/api/report";
 import { useGetQuestionDetailQuery, useGetQuestionSummaryQuery } from "@/shared/api/question";
-import { QuestionRenderer } from "@/features/test-participate/ui/QuestionRenderer";
-import { FivesecAnswerPage } from "@/features/question-fivesec/answer/FivesecAnswerPage";
-
-import type { ParticipateQuestion } from "@/features/test-participate/model/types";
 
 interface Props {
   testId: string;
 }
 
-type BadgeConfig = {
-  text: string;
-  color: "green" | "elephant" | "red" | "yellow";
-  variant: "weak";
-};
-
-const STATUS_BADGE: Record<TestStatus, BadgeConfig> = {
-  WAITING: { text: "검토중", color: "yellow", variant: "weak" },
-  IN_PROGRESS: { text: "진행중", color: "green", variant: "weak" },
-  COMPLETED: { text: "종료", color: "elephant", variant: "weak" },
-  REJECTED: { text: "반려", color: "red", variant: "weak" },
-};
-
-const REVIEW_POLICY_ITEMS = [
-  { title: "디지털 자산 및 가상자산 관련 서비스", desc: "디지털 자산의 소유, 이전, 저장, 거래, 중개, 발행(NFT 포함) 등의 기능을 제공하는 서비스는 법적 요건 충족 여부와 관계없이 토스 플랫폼에서는 자산 손실, 소비자 피해, 자금세탁 리스크에 따라 등록이 불가해요." },
-  { title: "자금세탁 가능성이 있는 서비스", desc: "미니앱 내에서 현금 또는 유사 자산의 직접적인 교환, 전환, 환불 기능이 포함된 경우, 거래 구조상 자금세탁 통로로 악용될 수 있기 때문에 등록이 불가해요." },
-  { title: "불법 또는 부정행위를 조장하는 서비스", desc: "법적으로 금지되거나 사회적 물의를 일으킬 수 있는 신분 조작, 해킹, 불법 문서 제공, 정보 수집 우회 등의 기능이 포함된 서비스는 명백히 등록이 불가해요." },
-  { title: "사행성 및 복권/베팅성 콘텐츠 포함 서비스", desc: "사행성 요소가 포함된 콘텐츠는 사용자 재산상 손실, 중독 유발, 연령 제한 문제 등으로 위법 소지가 있으며, 사용자 보호 및 서비스 신뢰도 확보를 위해 등록이 불가해요." },
-  { title: "금융 상품 중개 · 판매 · 광고 서비스", desc: "대출, 보험, 카드, 증권 등 금융 상품 관련 서비스는 법적 인허가 여부와 관계없이 소비자 보호, 금융정보의 정확성, 오인 가능성 등으로 인한 운영 리스크를 방지하기 위해 등록이 불가능하며, 향후 내부 정책 및 기준 정비에 따라 오픈 여부가 검토될 수 있어요." },
-  { title: "투자 자문, 리딩방, 유료 정보 제공 서비스", desc: "특정 종목 추천이나 투자 전략 안내 등으로 개인 투자자의 의사결정에 영향을 미치는 서비스는 운영 리스크 및 정책적 수용 미비로 인해 등록이 불가해요." },
-  { title: "의료 관련 서비스", desc: "비대면 진료 제공 또는 연결, 의료 행위로의 직접적인 연결, 병원 예약 기능, 병원으로부터 광고비를 수취하는 구조(유저 유입 기반 수익 모델), 병원 홍보/마케팅으로 해석될 수 있는 수익 구조 등의 경우 서비스 출시가 불가해요." },
-  { title: "이외 내부 정책상 승인 불가 서비스", desc: "법률 위반 여부와 관계없이 토스의 브랜드 신뢰성, UX 정책, 리스크 관리 방침에 따라 등록이 제한될 수 있어요." },
-] as const;
-
-const OVERLAY_MOTION = {
-  initial: { opacity: 0 },
-  animate: { opacity: 1 },
-  exit: { opacity: 0 },
-  transition: { duration: 0.2 },
-} as const;
-
-function QuestionPreviewSkeleton({ onClose }: { onClose: () => void }) {
-  return (
-    <motion.div {...OVERLAY_MOTION} className="fixed inset-0 z-60 flex flex-col bg-white">
-      <div className="flex-1 overflow-y-auto px-5 pt-8">
-        <Skeleton custom={["title", "subtitle"]} repeatLastItemCount={0} background="greyOpacity100" />
-        <div className="mt-6">
-          <Skeleton custom={["card"]} repeatLastItemCount={0} height={200} background="greyOpacity100" />
-        </div>
-        <div className="mt-6">
-          <Skeleton custom={["list"]} repeatLastItemCount={3} background="greyOpacity100" />
-        </div>
-      </div>
-      <BottomCTA.Single color="dark" variant="weak" onClick={onClose}>
-        뒤로가기
-      </BottomCTA.Single>
-    </motion.div>
-  );
-}
-
-function QuestionPreviewOverlay({
-  question,
-  onClose,
-}: {
-  question: ParticipateQuestion;
-  onClose: () => void;
-}) {
-  return (
-    <motion.div {...OVERLAY_MOTION} className="fixed inset-0 z-60 flex flex-col bg-white">
-      <div className="flex-1 overflow-y-auto">
-        <QuestionRenderer
-          question={question}
-          answer={undefined}
-          onChange={() => {}}
-          onPrev={onClose}
-          onGoNext={onClose}
-          isFirst={true}
-          isLast={true}
-        />
-      </div>
-      <BottomCTA.Single color="dark" variant="weak" onClick={onClose}>
-        뒤로가기
-      </BottomCTA.Single>
-    </motion.div>
-  );
-}
-
-function FivesecPreviewOverlay({
-  question,
-  onClose,
-}: {
-  question: Extract<ParticipateQuestion, { type: "FIVE_SECOND" }>;
-  onClose: () => void;
-}) {
-  const [answer, setAnswer] = useState<{ type: "FIVE_SECOND"; selectedIds: string[] }>({
-    type: "FIVE_SECOND",
-    selectedIds: [],
-  });
-
-  return (
-    <motion.div
-      {...OVERLAY_MOTION}
-      className="fixed inset-0 z-49 flex flex-col overflow-y-auto bg-white pb-10"
-    >
-      <FivesecAnswerPage
-        question={question}
-        answer={answer}
-        onChange={setAnswer}
-        onPrev={onClose}
-        onGoNext={onClose}
-        isFirst={false}
-        isLast={true}
-        prevLabel="돌아가기"
-        isPreview={true}
-      />
-    </motion.div>
-  );
-}
-
 function TestResultSkeleton() {
   return (
     <div className="flex flex-col">
-      {/* Top 영역 */}
       <div className="px-5 pt-8 pb-4">
         <Skeleton
           custom={["subtitle", "title", "subtitle"]}
@@ -150,23 +35,12 @@ function TestResultSkeleton() {
           background="greyOpacity100"
         />
       </div>
-
-      {/* 버튼 영역 */}
       <div className="px-5 pb-3">
-        <Skeleton
-          custom={["card"]}
-          repeatLastItemCount={0}
-          height={52}
-          background="greyOpacity100"
-        />
+        <Skeleton custom={["card"]} repeatLastItemCount={0} height={52} background="greyOpacity100" />
       </div>
-
-      {/* 탭 영역 */}
       <div className="px-5 pb-4">
         <Skeleton custom={["subtitle"]} repeatLastItemCount={0} background="greyOpacity100" />
       </div>
-
-      {/* 질문 목록 영역 */}
       <div className="px-5">
         <Skeleton custom={["listWithIcon"]} repeatLastItemCount={4} background="greyOpacity100" />
       </div>
@@ -179,25 +53,17 @@ export function TestResultPage({ testId }: Props) {
   const [isDownloadSheetOpen, setIsDownloadSheetOpen] = useState(false);
   const [selectedQuestionId, setSelectedQuestionId] = useState<number | null>(null);
 
-  const [isGuideExpanded, setIsGuideExpanded] = useState(false);
-
   const { data: reportData, isLoading } = useGetReportQuery(Number(testId));
   const report = reportData?.data;
 
   const { data: questionSummary = [] } = useGetQuestionSummaryQuery(Number(testId));
-
-  const {
-    data: previewQuestion,
-    isLoading: isQuestionLoading,
-  } = useGetQuestionDetailQuery(Number(testId), selectedQuestionId);
+  const { data: previewQuestion } = useGetQuestionDetailQuery(Number(testId), selectedQuestionId);
 
   const testStatus: TestStatus = report?.testStatus ?? "IN_PROGRESS";
   const isEnded = testStatus === "COMPLETED";
   const isReviewState = testStatus === "WAITING" || testStatus === "REJECTED";
   const showParticipant = testStatus === "IN_PROGRESS" || testStatus === "COMPLETED";
   const results = (report?.reports ?? []).map(mapReportItemToQuestionResult);
-
-  const isFivesecPreview = previewQuestion?.type === "FIVE_SECOND";
 
   if (isLoading) {
     return <TestResultSkeleton />;
@@ -239,92 +105,20 @@ export function TestResultPage({ testId }: Props) {
 
       {isReviewState ? (
         <>
-          <ListRow
-            left={<ListRow.AssetIcon size="xsmall" shape="original" name="icn-service-316-color" />}
-            contents={
-              <ListRow.Texts
-                type="1RowTypeC"
-                top="내부 검토 기준 확인하기"
-                topProps={{ color: adaptive.grey800 }}
-              />
-            }
-            verticalPadding="xlarge"
-            arrowType={isGuideExpanded ? "up" : "down"}
-            onClick={() => setIsGuideExpanded((prev) => !prev)}
-          />
-          <AnimatePresence initial={false}>
-            {isGuideExpanded && (
-              <motion.div
-                initial={{ height: 0, opacity: 0 }}
-                animate={{ height: "auto", opacity: 1 }}
-                exit={{ height: 0, opacity: 0 }}
-                transition={{ duration: 0.25, ease: "easeInOut" }}
-                style={{ overflow: "hidden" }}
-              >
-                <div
-                  style={{
-                    width: "100%",
-                    backgroundColor:
-                      "var(--token-tds-color-grey-background, var(--adaptiveGreyBackground, #f2f4f6))",
-                    paddingTop: 16,
-                    paddingBottom: 16,
-                    paddingLeft: 8,
-                    paddingRight: 8,
-                  }}
-                >
-                  <Post.Paragraph paddingBottom={4} typography="t6">
-                    앱인토스 필수 정책에 따라 해당 테스트가 반려 되었어요.
-                  </Post.Paragraph>
-                  <Post.Paragraph paddingBottom={16} typography="t6">
-                    아래에 해당하는 서비스 테스트는 등록이 불가해요.
-                  </Post.Paragraph>
-
-                  {REVIEW_POLICY_ITEMS.map(({ title, desc }, i) => (
-                    <div key={i}>
-                      <Post.Paragraph paddingBottom={2} typography="t6" fontWeight="bold">
-                        {`${i + 1}. ${title}`}
-                      </Post.Paragraph>
-                      <Post.Paragraph paddingBottom={i < 7 ? 12 : 0} typography="t6">
-                        {desc}
-                      </Post.Paragraph>
-                    </div>
-                  ))}
-
-                  <Post.Hr paddingBottom={0} />
-
-                  <Post.Ul paddingBottom={0} typography="t6">
-                    <Post.Li>
-                      심사 결과는 내부 정책 및 리스크 검토 절차에 따라 변경될 수 있으며, 사전 등록을
-                      보장하지 않아요.
-                    </Post.Li>
-                    <Post.Li>
-                      서비스 특수성, 비즈니스 모델에 따라 사전 상담 또는 추가 설명 요청이 있을 수
-                      있어요.
-                    </Post.Li>
-                    <Post.Li>
-                      기존 회사 및 서비스를 단순 홍보하기 위해서 앱인토스 미니앱을 출시할 수 없어요.
-                    </Post.Li>
-                    <Post.Li>
-                      위 내용 외에도 사용자 보호 및 신뢰성 확보를 위한 추가 기준이 적용될 수 있어요.
-                    </Post.Li>
-                  </Post.Ul>
-                </div>
-              </motion.div>
-            )}
-          </AnimatePresence>
+          <ReviewGuideAccordion />
           <ListHeader
             titleWidthRatio={0.6}
             title={
-              <ListHeader.TitleParagraph
-                typography="t5"
-                fontWeight="medium"
-                color={adaptive.grey800}
-              >
+              <ListHeader.TitleParagraph typography="t5" fontWeight="medium" color={adaptive.grey800}>
                 질문 목록
               </ListHeader.TitleParagraph>
             }
           />
-          <QuestionTabContent questions={questionSummary} onSelectQuestion={setSelectedQuestionId} noPadding />
+          <QuestionTabContent
+            questions={questionSummary}
+            onSelectQuestion={setSelectedQuestionId}
+            noPadding
+          />
         </>
       ) : (
         <>
@@ -354,7 +148,12 @@ export function TestResultPage({ testId }: Props) {
             </Tab.Item>
           </Tab>
 
-          {selectedTabIndex === 0 && <QuestionTabContent questions={questionSummary} onSelectQuestion={setSelectedQuestionId} />}
+          {selectedTabIndex === 0 && (
+            <QuestionTabContent
+              questions={questionSummary}
+              onSelectQuestion={setSelectedQuestionId}
+            />
+          )}
 
           {selectedTabIndex === 1 && !isEnded && (
             <Result
@@ -375,18 +174,11 @@ export function TestResultPage({ testId }: Props) {
         </>
       )}
 
-      {selectedQuestionId !== null && isQuestionLoading && (
-        <QuestionPreviewSkeleton onClose={() => setSelectedQuestionId(null)} />
-      )}
-      {previewQuestion != null && !isFivesecPreview && (
-        <QuestionPreviewOverlay question={previewQuestion} onClose={() => setSelectedQuestionId(null)} />
-      )}
-      {isFivesecPreview && previewQuestion != null && (
-        <FivesecPreviewOverlay
-          question={previewQuestion as Extract<ParticipateQuestion, { type: "FIVE_SECOND" }>}
-          onClose={() => setSelectedQuestionId(null)}
-        />
-      )}
+      <QuestionPreviewOverlay
+        selectedQuestionId={selectedQuestionId}
+        previewQuestion={previewQuestion}
+        onClose={() => setSelectedQuestionId(null)}
+      />
 
       <DownloadSheet open={isDownloadSheetOpen} onClose={() => setIsDownloadSheetOpen(false)} />
     </div>
